fix(dhcp): address Copilot review — 5 correctness fixes
1. server: handle_request short-circuit bug — is_available() accepted
   the IP without marking it allocated; now always uses
   allocate_specific() which atomically checks + allocates.

2. server: reserve_ip now asserts the IP is available before inserting,
   preventing silent IP collisions from out-of-range or duplicate
   reservations.

3. packet: DNS option serialization caps at 63 servers and resizes the
   buffer dynamically, preventing out-of-bounds panic on large lists.

4. config: DhcpConfig::new derives pool range from subnet/netmask with
   saturating arithmetic instead of unchecked server_ip+1/+253, and
   skips server_ip to avoid assigning the gateway.

5. allocator: range_size and available_count use u64 arithmetic to
   avoid u32 overflow on large IP ranges.

+++ crates/arcbox-dhcp/src/allocator.rs
@@ -47,8 +47,7 @@ impl IpAllocator {
     ///
     /// Returns `None` if no addresses are available.
     pub fn allocate(&mut self) -> Option<Ipv4Addr> {
-        // Try sequential allocation first (faster for sparse usage)
-        let range_size = self.end - self.start + 1;
+        let range_size = u64::from(self.end) - u64::from(self.start) + 1;
 
         for _ in 0..range_size {
             if !self.allocated.contains(&self.next) {
@@ -113,7 +112,7 @@ impl IpAllocator {
     /// Returns the number of available addresses.
     #[must_use]
     pub fn available_count(&self) -> usize {
-        let total = (self.end - self.start + 1) as usize;
+        let total = (u64::from(self.end) - u64::from(self.start) + 1) as usize;
         total - self.allocated.len()
     }
 }
